GVT-1677 Move selection panel geometry sectionyy into a slightly separated component
