+++ ui/src/selection-panel/selection-panel.tsx
@@ -9,25 +9,16 @@ import {
     LayoutTrackNumber,
     LayoutTrackNumberId,
 } from 'track-layout/track-layout-model';
-import { GeometryPlanPanel } from 'selection-panel/geometry-plan-panel/geometry-plan-panel';
 import {
     OnSelectOptions,
     OpenedPlanLayout,
     OptionalItemCollections,
     SelectableItemType,
 } from 'selection/selection-model';
-import {
-    GeometryPlanHeader,
-    GeometryPlanId,
-    SortByValue,
-    SortOrderValue,
-} from 'geometry/geometry-model';
-import { createClassName } from 'vayla-design-lib/utils';
 import { KmPostsPanel } from 'selection-panel/km-posts-panel/km-posts-panel';
 import SwitchPanel from 'selection-panel/switch-panel/switch-panel';
 import { getTrackNumbers } from 'track-layout/layout-track-number-api';
 import TrackNumberPanel from 'selection-panel/track-number-panel/track-number-panel';
-import { getGeometryPlanHeaders, getTrackLayoutPlan } from 'geometry/geometry-api';
 import { MapViewport } from 'map/map-model';
 import {
     createEmptyItemCollections,
@@ -42,10 +33,7 @@ import { PublishType } from 'common/common-model';
 import { useTranslation } from 'react-i18next';
 import { LocationTracksPanel } from 'selection-panel/alignment-panel/location-tracks-panel';
 import ReferenceLinesPanel from 'selection-panel/alignment-panel/reference-lines-panel';
-import { Eye } from 'geoviite-design-lib/eye/eye';
-import { GeometryPlanLinkStatus } from 'linking/linking-model';
-import { getPlanLinkStatus } from 'linking/linking-api';
-import { useMapState, useSetState } from 'utils/react-utils';
+import SelectionPanelGeometrySection from './selection-panel-geometry-section';
 
 type SelectionPanelProps = {
     changeTimes: ChangeTimes;
@@ -68,13 +56,6 @@ type SelectionPanelProps = {
     togglePlanKmPostsOpen: (payload: ToggleAccordionOpenPayload) => void;
     togglePlanAlignmentsOpen: (payload: ToggleAccordionOpenPayload) => void;
     togglePlanSwitchesOpen: (payload: ToggleAccordionOpenPayload) => void;
-};
-
-const MAX_PLAN_HEADERS = 50;
-
-type LoadedGeometryPlan = {
-    planLayout: GeometryPlanLayout;
-    linkStatus: GeometryPlanLinkStatus;
 };
 
 const SelectionPanel: React.FC<SelectionPanelProps> = ({
@@ -102,38 +83,6 @@ const SelectionPanel: React.FC<SelectionPanelProps> = ({
     const { t } = useTranslation();
     const [visibleTrackNumbers, setVisibleTrackNumbers] = React.useState<LayoutTrackNumber[]>([]);
     const [trackNumberFilter, setTrackNumberFilter] = React.useState<LayoutTrackNumber[]>([]);
-    const [planHeaders, setPlanHeaders] = React.useState<GeometryPlanHeader[]>([]);
-    const [loadedPlans, setLoadedPlan] = useMapState<GeometryPlanId, LoadedGeometryPlan>();
-    const [plansBeingLoaded, startLoadingPlan, finishLoadingPlan] = useSetState<GeometryPlanId>();
-    const [planHeaderCount, setPlanHeaderCount] = React.useState<number>(0);
-
-    React.useEffect(() => {
-        getGeometryPlanHeaders(
-            MAX_PLAN_HEADERS,
-            0,
-            viewport.area,
-            ['GEOVIITE', 'GEOMETRIAPALVELU', 'PAIKANNUSPALVELU'],
-            trackNumberFilter.map((tn) => tn.id),
-            undefined,
-            SortByValue.UPLOADED_AT,
-            SortOrderValue.ASCENDING,
-        ).then((page) => {
-            setPlanHeaders(page.items);
-            setPlanHeaderCount(page.totalCount);
-        });
-    }, [viewport.area, changeTimes.geometryPlan, trackNumberFilter]);
-
-    React.useEffect(
-        () => [...loadedPlans.keys()].forEach(loadPlanLayout),
-        [
-            publishType,
-            changeTimes.geometryPlan,
-            changeTimes.layoutReferenceLine,
-            changeTimes.layoutLocationTrack,
-            changeTimes.layoutSwitch,
-            changeTimes.layoutKmPost,
-        ],
-    );
 
     const toggleTrackNumberFilter = React.useCallback(
         (tn: LayoutTrackNumber) => {
@@ -207,41 +156,6 @@ const SelectionPanel: React.FC<SelectionPanelProps> = ({
     }, [referenceLines, trackNumberFilter]);
 
     const filteredKmPosts = kmPosts.filter((km) => filterByTrackNumberId(km.trackNumberId));
-    const planHeaderIds = planHeaders
-        .map((plan) => plan.id)
-        .reduce((set, id) => set.add(id), new Set());
-    const visibleSelectedPlans = (selectedPlanLayouts ?? []).filter((plan) =>
-        planHeaderIds.has(plan.planId),
-    );
-    const anyPlansVisible = visibleSelectedPlans.length > 0;
-
-    const loadPlanLayout = (id: GeometryPlanId) => {
-        startLoadingPlan(id);
-        const rv = Promise.all([
-            getTrackLayoutPlan(id, changeTimes.geometryPlan, false),
-            getPlanLinkStatus(id, publishType),
-        ]).then(([planLayout, linkStatus]) => {
-            if (planLayout) {
-                setLoadedPlan(id, { planLayout, linkStatus });
-            }
-            return planLayout;
-        });
-        rv.finally(() => finishLoadingPlan(id));
-        return rv;
-    };
-
-    const toggleAllPlanVisibilities = () => {
-        if (anyPlansVisible) {
-            visibleSelectedPlans.forEach(onTogglePlanVisibility);
-        } else {
-            planHeaders.forEach((ph) => {
-                loadPlanLayout(ph.id).then((loadedPlan) => {
-                    onTogglePlanVisibility(loadedPlan);
-                });
-            });
-        }
-    };
-
     return (
         <div className={styles['selection-panel']}>
             <section>
@@ -256,107 +170,24 @@ const SelectionPanel: React.FC<SelectionPanelProps> = ({
                     />
                 </div>
             </section>
-            {planHeaders && (
-                <section>
-                    <h3 className={styles['selection-panel__title']}>
-                        {`${t('selection-panel.geometries-title')} (${
-                            planHeaders.length
-                        }/${planHeaderCount})`}{' '}
-                        {planHeaders.length > 1 && planHeaders.length === planHeaderCount && (
-                            <Eye
-                                onVisibilityToggle={toggleAllPlanVisibilities}
-                                visibility={anyPlansVisible}
-                            />
-                        )}
-                    </h3>
-                    <div
-                        className={createClassName(
-                            styles['selection-panel__content'],
-                            styles['selection-panel__content--unpadded'],
-                        )}>
-                        {planHeaders.length == planHeaderCount &&
-                            planHeaders.map((h) => {
-                                return (
-                                    <GeometryPlanPanel
-                                        key={h.id}
-                                        planHeader={h}
-                                        onPlanHeaderSelection={(header) =>
-                                            onSelect({
-                                                ...createEmptyItemCollections(),
-                                                geometryPlans: [header],
-                                                isToggle: true,
-                                            })
-                                        }
-                                        publishType={publishType}
-                                        changeTimes={changeTimes}
-                                        onTogglePlanVisibility={onTogglePlanVisibility}
-                                        onToggleAlignmentVisibility={onToggleAlignmentVisibility}
-                                        onToggleAlignmentSelection={(alignment) =>
-                                            onSelect({
-                                                ...createEmptyItemCollections(),
-                                                geometryAlignments: [
-                                                    {
-                                                        geometryItem: alignment,
-                                                        planId: h.id,
-                                                    },
-                                                ],
-                                                isToggle: true,
-                                            })
-                                        }
-                                        onToggleSwitchVisibility={onToggleSwitchVisibility}
-                                        onToggleSwitchSelection={(switchItem) =>
-                                            onSelect({
-                                                ...createEmptyItemCollections(),
-                                                geometrySwitches: [
-                                                    {
-                                                        geometryItem: switchItem,
-                                                        planId: h.id,
-                                                    },
-                                                ],
-                                                isToggle: true,
-                                            })
-                                        }
-                                        onToggleKmPostVisibility={onToggleKmPostVisibility}
-                                        onToggleKmPostSelection={(kmPost) =>
-                                            onSelect({
-                                                ...createEmptyItemCollections(),
-                                                geometryKmPosts: [
-                                                    {
-                                                        geometryItem: kmPost,
-                                                        planId: h.id,
-                                                    },
-                                                ],
-                                                isToggle: true,
-                                            })
-                                        }
-                                        selectedItems={selectedItems}
-                                        selectedPlanLayouts={selectedPlanLayouts}
-                                        togglePlanOpen={togglePlanOpen}
-                                        openedPlanLayouts={openedPlanLayouts}
-                                        togglePlanKmPostsOpen={togglePlanKmPostsOpen}
-                                        togglePlanAlignmentsOpen={togglePlanAlignmentsOpen}
-                                        togglePlanSwitchesOpen={togglePlanSwitchesOpen}
-                                        planLayout={loadedPlans.get(h.id)?.planLayout ?? null}
-                                        linkStatus={loadedPlans.get(h.id)?.linkStatus ?? null}
-                                        planBeingLoaded={plansBeingLoaded.has(h.id)}
-                                        loadPlanLayout={() => loadPlanLayout(h.id)}
-                                    />
-                                );
-                            })}
-                        {planHeaders.length < planHeaderCount && (
-                            <span className={styles['selection-panel__subtitle']}>{`${t(
-                                'selection-panel.zoom-closer',
-                            )}`}</span>
-                        )}
-
-                        {planHeaders.length === 0 && (
-                            <span className={styles['selection-panel__subtitle']}>
-                                {`${t('selection-panel.no-results')}`}{' '}
-                            </span>
-                        )}
-                    </div>
-                </section>
-            )}
+            <SelectionPanelGeometrySection
+                publishType={publishType}
+                changeTimes={changeTimes}
+                selectedItems={selectedItems}
+                selectedPlanLayouts={selectedPlanLayouts}
+                viewport={viewport}
+                onToggleAlignmentVisibility={onToggleAlignmentVisibility}
+                onToggleKmPostVisibility={onToggleKmPostVisibility}
+                onTogglePlanVisibility={onTogglePlanVisibility}
+                onToggleSwitchVisibility={onToggleSwitchVisibility}
+                openedPlanLayouts={openedPlanLayouts}
+                togglePlanKmPostsOpen={togglePlanKmPostsOpen}
+                togglePlanAlignmentsOpen={togglePlanAlignmentsOpen}
+                togglePlanSwitchesOpen={togglePlanSwitchesOpen}
+                trackNumberFilter={trackNumberFilter}
+                togglePlanOpen={togglePlanOpen}
+                onSelect={onSelect}
+            />
             <section>
                 <h3 className={styles['selection-panel__title']}>
                     {t('selection-panel.km-posts-title')} ({filteredKmPosts.length}/{kmPosts.length}
